@@ -1,9 +1,10 @@
 package cn.mcmod.arsenal;
 
-import net.minecraftforge.common.ForgeConfigSpec;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.config.ModConfigEvent;
+
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.event.config.ModConfigEvent;
+import net.neoforged.neoforge.common.ModConfigSpec;
 
 @Mod.EventBusSubscriber(modid = "arsenal_core", bus = Mod.EventBusSubscriber.Bus.MOD)
 public class ArsenalConfig {
@@ -11,23 +12,23 @@ public class ArsenalConfig {
     public static boolean normal_sword_foil;
     public static boolean xuanyuanjian_foil;
 
-    private static final ForgeConfigSpec.Builder BUILDER = new ForgeConfigSpec.Builder()
+    private static final ModConfigSpec.Builder BUILDER = new ModConfigSpec.Builder()
             .comment("Arsenal Core Config")
             .push("General");
 
-    private static final ForgeConfigSpec.DoubleValue MAXIMUM_POWER_DAMAGE = BUILDER
+    private static final ModConfigSpec.DoubleValue MAXIMUM_POWER_DAMAGE = BUILDER
             .comment("How high is the damage value of xuan yuan jian's power")
             .defineInRange("maximum_power_damage", 25.0, 0.0, 1000.0);
 
-    private static final ForgeConfigSpec.BooleanValue NORMAL_SWORD_FOIL = BUILDER
+    private static final ModConfigSpec.BooleanValue NORMAL_SWORD_FOIL = BUILDER
             .comment("Is normal swords foil?")
             .define("foil_normal", true);
 
-    private static final ForgeConfigSpec.BooleanValue XUANYUANJIAN_FOIL = BUILDER
+    private static final ModConfigSpec.BooleanValue XUANYUANJIAN_FOIL = BUILDER
             .comment("Is the Xuan Yuan Jian foil? Not recommended")
             .define("foil_xuanyuanjian", false);
 
-    public static final ForgeConfigSpec SPEC = BUILDER.pop().build();
+    public static final ModConfigSpec SPEC = BUILDER.pop().build();
 
     @SubscribeEvent
     static void onLoad(final ModConfigEvent event) {
