@@ -5,29 +5,23 @@
 
 package cn.mcmod.arsenal;
 
-import cn.mcmod.arsenal.api.tier.WeaponTier;
 import cn.mcmod.arsenal.api.tier.WeaponTierRegistry;
+import cn.mcmod.arsenal.compat.curios.CuriosCapProvider;
 import cn.mcmod.arsenal.item.ArsenalCreativeModTab;
 import cn.mcmod.arsenal.item.ItemRegistry;
-import cn.mcmod.arsenal.net.NetPacketHandler;
+import cn.mcmod.arsenal.net.DrawSwordPacket;
 import cn.mcmod.arsenal.tier.VanillaWeaponTiers;
-import net.minecraft.core.registries.Registries;
-import net.minecraft.network.chat.Component;
-import net.minecraft.world.item.CreativeModeTab;
-import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.fml.ModList;
-import net.minecraftforge.fml.ModLoadingContext;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.config.ModConfig;
-import net.minecraftforge.fml.config.ModConfig.Type;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.registries.DeferredRegister;
-import net.minecraftforge.registries.RegistryObject;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.ModList;
+import net.neoforged.fml.ModLoadingContext;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.config.ModConfig;
+import net.neoforged.neoforge.network.event.RegisterPayloadHandlerEvent;
+import net.neoforged.neoforge.network.registration.IPayloadRegistrar;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.util.function.Supplier;
 
 @Mod(ArsenalCore.MODID)
 public class ArsenalCore {
@@ -35,15 +29,25 @@ public class ArsenalCore {
     public static boolean curiosLoaded = false;
     private static final Logger LOGGER = LogManager.getLogger();
 
-    public ArsenalCore() {
+    public ArsenalCore(IEventBus modEventBus) {
         curiosLoaded = ModList.get().isLoaded("curios");
         this.registerVanillaTiers();
-        FMLJavaModLoadingContext.get().getModEventBus().addListener(this::setup);
-        ItemRegistry.ITEMS.register(FMLJavaModLoadingContext.get().getModEventBus());
-        ArsenalCreativeModTab.CREATIVE_MODE_TABS.register(FMLJavaModLoadingContext.get().getModEventBus());
+        modEventBus.addListener(this::registerPayloadHandlers);
+        ItemRegistry.ITEMS.register(modEventBus);
+        CuriosCapProvider.register(modEventBus);
+        ArsenalCreativeModTab.CREATIVE_MODE_TABS.register(modEventBus);
         ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, ArsenalConfig.SPEC);
     }
 
+    @SubscribeEvent
+    public void registerPayloadHandlers (final RegisterPayloadHandlerEvent event) {
+        final IPayloadRegistrar registrar = event.registrar(MODID)
+                .versioned("1.0");
+
+        registrar.play(DrawSwordPacket.ID, DrawSwordPacket::new, handler -> handler
+                .server(DrawSwordPacket::handle)
+        );
+    }
     private void registerVanillaTiers() {
         WeaponTierRegistry.registerAll(
                 VanillaWeaponTiers.WOOD.get(),
@@ -55,9 +59,6 @@ public class ArsenalCore {
                 VanillaWeaponTiers.MAXIMUM_POWER.get());
     }
 
-    private void setup(FMLCommonSetupEvent event) {
-        event.enqueueWork(NetPacketHandler::registerMessage);
-    }
 
     public static Logger getLogger() {
         return LOGGER;
