@@ -7,24 +7,29 @@ package cn.mcmod.arsenal.client;
 
 import cn.mcmod.arsenal.ArsenalCore;
 import cn.mcmod.arsenal.net.DrawSwordPacket;
-import cn.mcmod.arsenal.net.NetPacketHandler;
 import com.mojang.blaze3d.platform.InputConstants.Type;
 import net.minecraft.client.KeyMapping;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.client.event.InputEvent;
-import net.minecraftforge.client.settings.KeyConflictContext;
-import net.minecraftforge.client.settings.KeyModifier;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
+import net.minecraft.client.Minecraft;
+import net.minecraft.network.protocol.common.ClientboundCustomPayloadPacket;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.neoforge.client.event.InputEvent;
+import net.neoforged.neoforge.client.settings.KeyConflictContext;
+import net.neoforged.neoforge.client.settings.KeyModifier;
 
-@EventBusSubscriber({Dist.CLIENT})
+@Mod.EventBusSubscriber({Dist.CLIENT})
 public class KeyDrawSword {
     public static final KeyMapping KEY = new KeyMapping("key.draw_sword", KeyConflictContext.IN_GAME, KeyModifier.CONTROL, Type.MOUSE, 1, "key.category.arsenal");
 
     @SubscribeEvent
     public static void onKeyboardInput(InputEvent.Key event) {
         if (KEY.isDown() && ArsenalCore.curiosLoaded) {
-            NetPacketHandler.INSTANCE.sendToServer(new DrawSwordPacket("Reika Battou"));
+            Minecraft mc = Minecraft.getInstance();
+            if (mc.player != null) {
+                mc.getConnection().send(new ClientboundCustomPayloadPacket(new DrawSwordPacket("Reika Battou")));
+            }
         }
     }
+
 }
