@@ -16,8 +16,8 @@ import net.minecraft.client.renderer.entity.RenderLayerParent;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemDisplayContext;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.api.distmarker.OnlyIn;
 import org.joml.Quaternionf;
 import top.theillusivec4.curios.api.SlotContext;
 import top.theillusivec4.curios.api.client.ICurioRenderer;
@@ -63,7 +63,7 @@ public class WeaponFrogRender implements ICurioRenderer {
                         false,
                         matrixStack,
                         renderTypeBuffer,
-                        livingEntity.level,
+                        livingEntity.level(),
                         light,
                         LivingEntityRenderer.getOverlayCoords(livingEntity, 0.0F),
                         livingEntity.getId());
