@@ -1,36 +1,36 @@
-//
-// Source code recreated from a .class file by IntelliJ IDEA
-// (powered by FernFlower decompiler)
-//
-
 package cn.mcmod.arsenal.item;
 
 import cn.mcmod.arsenal.ArsenalCore;
 import cn.mcmod.arsenal.compat.curios.CuriosCapProvider;
 import cn.mcmod.arsenal.compat.curios.ItemHandlerCapProvider;
-import net.minecraft.core.Direction;
-import net.minecraft.nbt.CompoundTag;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.common.capabilities.ForgeCapabilities;
-import net.minecraftforge.common.capabilities.ICapabilityProvider;
-import net.minecraftforge.common.util.NonNullSupplier;
-import net.minecraftforge.items.IItemHandler;
-import net.minecraftforge.items.ItemStackHandler;
+import net.minecraft.world.level.Level;
+import net.neoforged.neoforge.items.ItemStackHandler;
 
 public class WeaponFrogItem extends Item {
-    public static final NonNullSupplier<IllegalArgumentException> CAPABILITY_EXCEPTION = () -> new IllegalArgumentException("Capability must not be null!");
-
     public WeaponFrogItem() {
         super(new Properties().stacksTo(1));
     }
 
     @Override
-    public ICapabilityProvider initCapabilities(ItemStack stack, CompoundTag nbt) {
-        return ArsenalCore.curiosLoaded ? new CuriosCapProvider(stack, nbt) : new ItemHandlerCapProvider(stack, nbt);
+    public void onCraftedBy(ItemStack stack, Level level, Player player) {
+        super.onCraftedBy(stack, level, player);
+        initAttachments(stack);
+    }
+
+    private void initAttachments(ItemStack stack) {
+        if (!stack.hasData(ItemHandlerCapProvider.ITEM_HANDLER)) {
+            stack.setData(ItemHandlerCapProvider.ITEM_HANDLER, new ItemStackHandler(1));
+        }
+
+        if (ArsenalCore.curiosLoaded) {
+            CuriosCapProvider.attachCurio(stack);
+        }
     }
 
     public static ItemStackHandler getInventory(ItemStack stack) {
-        return (ItemStackHandler) stack.getCapability(ForgeCapabilities.ITEM_HANDLER, null).orElseThrow(CAPABILITY_EXCEPTION);
+        return ItemHandlerCapProvider.getInventory(stack);
     }
 }
