@@ -20,14 +20,13 @@ import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.SwordItem;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.UseAnim;
 import net.minecraft.world.level.Level;
-import net.minecraftforge.common.ToolAction;
-import net.minecraftforge.common.ToolActions;
+import net.neoforged.neoforge.common.ToolAction;
+import net.neoforged.neoforge.common.ToolActions;
 
 public class ChineseSwordItem extends SwordItem implements IDrawable, IWeaponTiered {
     private final WeaponTier tier;
@@ -71,8 +70,8 @@ public class ChineseSwordItem extends SwordItem implements IDrawable, IWeaponTie
 
 
     @Override
-    public boolean isFoil(ItemStack p_77636_1_) {
-        return ArsenalConfig.normal_sword_foil && super.isFoil(p_77636_1_);
+    public boolean isFoil(ItemStack stack) {
+        return ArsenalConfig.normal_sword_foil && super.isFoil(stack);
     }
 
     @Override
