@@ -15,10 +15,9 @@ import cn.mcmod.arsenal.util.Lazy;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.world.item.Tiers;
-import net.minecraftforge.common.Tags.Items;
+import net.neoforged.neoforge.common.Tags;
 
 public final class VanillaWeaponTiers {
-    // 将字段类型改为 Supplier<WeaponTier> 并直接初始化
     public static final Supplier<WeaponTier> WOOD = new Lazy<>(() -> new BlankTier("wooden", Tiers.WOOD, ItemTags.PLANKS.location()));
     public static final Supplier<WeaponTier> STONE = new Lazy<>(() -> new BlankTier("stone", Tiers.STONE, new ResourceLocation("forge:cobblestone")));
     public static final Supplier<WeaponTier> IRON = new Lazy<>(() -> new BlankTier("iron", Tiers.IRON, new ResourceLocation("forge:ingots/iron")));
@@ -26,7 +25,7 @@ public final class VanillaWeaponTiers {
     public static final Supplier<WeaponTier> DIAMOND = new Lazy<>(() -> new BlankTier("diamond", Tiers.DIAMOND, new ResourceLocation("forge:gems/diamond")));
     public static final Supplier<WeaponTier> NETHERITE = new Lazy<>(() -> new BlankTier("netherite", Tiers.NETHERITE, new ResourceLocation("forge:ingots/netherite")));
     public static final Supplier<WeaponTier> COPPER = new Lazy<>(() -> new BlankTier("copper", "arsenal_core", 1, 200, 5.0F, 1.5F, 8, new ResourceLocation("forge:ingots/copper")));
-    public static final Supplier<WeaponTier> LAPIS = new Lazy<>(() -> new BlankTier("lapis_lazuli", "arsenal_core", 2, 200, 2.0F, 2.0F, 40, Items.GEMS_LAPIS.location()));
+    public static final Supplier<WeaponTier> LAPIS = new Lazy<>(() -> new BlankTier("lapis_lazuli", "arsenal_core", 2, 200, 2.0F, 2.0F, 40, Tags.Items.GEMS_LAPIS.location()));
     public static final Supplier<WeaponTier> MAXIMUM_POWER = new Lazy<>(() -> (new WeaponTier("maximum_power", "arsenal_core", 5, -1, 8.0F, (float) ArsenalConfig.maximum_power_damage, 50, new ResourceLocation("forge:gems/diamond"), new XuanyuanFeature())).setSpecial());
 
 }
